The verification version is long now
To be uniform with the other consensus parameters

+++ io-hotmoka-beans/src/main/java/io/hotmoka/beans/responses/TransactionResponseWithInstrumentedJar.java
@@ -51,5 +51,5 @@ public interface TransactionResponseWithInstrumentedJar {
 	 * 
 	 * @return the version
 	 */
-	int getVerificationVersion();
+	long getVerificationVersion();
 }
