+++ io-hotmoka-local/src/main/java/io/hotmoka/local/internal/EngineClassLoaderImpl.java
@@ -821,7 +821,7 @@ public final class EngineClassLoaderImpl implements EngineClassLoader {
 	}
 
 	@Override
-	public int getVerificationVersion() {
+	public long getVerificationVersion() {
 		return parent.getVerificationVersion();
 	}
 }
